Resuming work on the game! Getting it to run again Updating to comply with the framework changes. Nothing beyond simple regex replacements.

+++ dbmodel.sql
@@ -31,7 +31,7 @@
 
 
 -- Example 2: add a custom field to the standard "player" table
-ALTER TABLE `player` ADD `custom_order` SMALLINT UNSIGNED NOT NULL DEFAULT '0';
+ALTER TABLE `player` ADD `custom_order` INT UNSIGNED NOT NULL DEFAULT '0';
 ALTER TABLE `player` ADD `dial_value` varchar(10) NOT NULL DEFAULT 'water';
 ALTER TABLE `player` ADD `dial_location` varchar(12) NOT NULL DEFAULT 'player';
